# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Library Management API",
      "description": "A RESTful API for managing a collection of books in a library."
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
      "/books": {
        "get": {
          "summary": "List all books",
          "description": "Returns a list of all books in the library.",
          "responses": {
            "200": {
              "description": "A list of books",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Book"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Add a new book",
          "description": "Add a book to the library with the required details.",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Book"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Book added successfully."
            },
            "400": {
              "description": "Missing required fields."
            }
          }
        }
      },
      "/books/search": {
        "get": {
          "summary": "Search for books",
          "description": "Filter books by author, published year, or genre.",
          "parameters": [
            {
              "name": "author",
              "in": "query",
              "type": "string",
              "description": "Filter by author."
            },
            {
              "name": "published_year",
              "in": "query",
              "type": "integer",
              "description": "Filter by published year."
            },
            {
              "name": "genre",
              "in": "query",
              "type": "string",
              "description": "Filter by genre."
            }
          ],
          "responses": {
            "200": {
              "description": "Filtered list of books.",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Book"
                }
              }
            }
          }
        }
      },
      "/books/{isbn}": {
        "delete": {
          "summary": "Delete a book",
          "description": "Remove a book from the library using its ISBN.",
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "The ISBN of the book to delete."
            }
          ],
          "responses": {
            "200": {
              "description": "Book deleted successfully."
            },
            "404": {
              "description": "Book not found."
            }
          }
        },
        "put": {
          "summary": "Update a book",
          "description": "Update the details of an existing book using its ISBN.",
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "The ISBN of the book to update."
            },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/BookUpdate"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Book updated successfully."
            },
            "404": {
              "description": "Book not found."
            }
          }
        }
      }
    },
    "definitions": {
      "Book": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "published_year": {
            "type": "integer"
          },
          "isbn": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          }
        },
        "required": ["title", "author", "published_year", "isbn"]
      },
      "BookUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "published_year": {
            "type": "integer"
          },
          "genre": {
            "type": "string"
          }
        }
      }
    }
  }
  